moved plotting functions to python/utils/plot_utils.py. Minor improvements to docs and logging
