+++ benchmark/repeat_finder_with_interruptions/run_repeat_finder.sh
@@ -9,8 +9,7 @@ set -ex
 # -i input sequence
 
 gunzip -c ../chr22.fa.gz > chr22.fa
-#time python3 ../../python/repeat_finder.py  --min-motif-size 1 --max-motif-size 6 --min-repeats 3 --min-span 9  --max-interruptions 1  -o chr22_repeats chr22.fa --show-progress-bar
-time python3 ../../python/repeat_finder.py  --min-motif-size 1 --max-motif-size 6 --min-repeats 3 --min-span 9  -i chr22:15226420-15229480  --plot repeats.svg ../chr22.fa.gz --show-progress-bar  --max-interruptions 1 -o chr22_repeats
+time python3 ../../python/repeat_finder.py  --min-motif-size 1 --max-motif-size 6 --min-repeats 3 --min-span 9  --max-interruptions 1  -o chr22_repeats chr22.fa --show-progress-bar
 rm chr22.fa
 
 set +x
@@ -18,5 +17,7 @@ echo output:
 wc -l chr22_repeats.bed
 
 exit 0
+
+time python3 ../../python/repeat_finder.py  --min-motif-size 1 --max-motif-size 6 --min-repeats 3 --min-span 9  -i chr22:15226420-15229480  --plot repeats.svg ../chr22.fa.gz --show-progress-bar  --max-interruptions 1 -o chr22_repeats
